% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours_to_pattern.R
\name{colours_to_pattern}
\alias{colours_to_pattern}
\title{Visualise the output from scaled_to_colours() into a cross-stitch pattern
#'}
\arguments{
\item{mosaic_object}{List output from scaled_to_colours(). Contains element  \code{Img_dmc}.}

\item{style}{Options for visualisation. Defaults to a cross-stitch pattern with DMC numbers}

\item{colour_list}{Whether a list of colours should also be output. Defaults to \code{TRUE}.}

\item{file_name}{Option to name the outputs. Defaults to \code{"embroidr_pattern"}}
}
\value{
An SVG files with the cross-stitch pattern. If \code{colour_list} is set to \code{TRUE},
a second SVG file containing a list of DMC colours used is also output.
}
\description{
Visualise the output from scaled_to_colours() into a cross-stitch pattern
#'
}
