% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pattern_from_image.R
\name{create_pattern_from_image}
\alias{create_pattern_from_image}
\title{Transform an image raster array into a cross-stitch pattern}
\arguments{
\item{image}{A raster array from an image.}

\item{size_unit}{The unit by which the image will be scaled. Can be one of the following options: 'stitches', 'inches', or 'cm'}

\item{img_size}{Size of the embroidery, in the unit determined by size_unit. Use a single value (e.g. \code{48}) for a square image.
Use a vector of two values for a rectangular image \code{c(width, height)}. NOTE: This parameter sets the size of the embroidery once
it will is embroidered. The embroidery pattern that will be outputted will be larger.}

\item{brightness}{A value >1 will increase the brightness of the image while a positive value <1 will decrease the brightness.}

\item{warhol}{Array of values \code{c(1, 2, 3)} associated with R, G, B colour channels. Swap values in array to swap colour channels for a fun visual effect.}

\item{method}{Default 'cie94'. The method to use for colour comparison. Either 'euclidean', 'cie1976', 'cie94', 'cie2000', or 'cmc'.
See \code{farver::compare_colour}.}

\item{colour_table}{Defaults to \code{dmc_colours}. Data frame of  colours to map onto image. Must contain name and red, green, blue channels.
See attached data  \code{dmc_colours} as examples. Original source for \code{dmc_colours} is: https://github.com/sharlagelfand/dmc}

\item{colour_palette}{Used to supply a character vector of specific DMC colours to use. Can be particularly useful in instances where you already have
many colours available and maintain a list of their DMC codes. Defaults to NULL - in which case all DMC colours are to be considered.
Ignored if a \code{colour_table} is supplied.}

\item{n_colours}{Optional: Integer representing the maximum number of colours to use. This option selects the \code{n_colours} that appear the most, and remaps all other colours to match these \code{n_colours}. Otherwise, defaults to NULL.}

\item{trans_bg}{If \code{img} is a png and has a transparent background, name of colour to replace the background.}

\item{dithering}{Improves colour of large, photo-realistic mosaics.}

\item{style}{Options for visualisation. Defaults to a tiled pattern with DMC numbers}

\item{colour_list}{Whether a list of colours should also be output. Defaults to \code{TRUE}.}

\item{file_name}{Option to name the output files. Defaults to \code{"embroidr_pattern"}}
}
\value{
An SVG file with the cross-stitch pattern. If \code{colour_list} is set to \code{TRUE},
a second SVG file containing a list of DMC colours used is also output.
}
\description{
Transform an image raster array into a cross-stitch pattern
}
