% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floss}
\alias{floss}
\title{DMC floss colours}
\format{
A tibble with 454 rows and 6 variables:
\describe{
\item{dmc}{dmc code}
\item{name}{dmc name}
\item{hex}{hex code of colour}
\item{red}{RGB colour code: red}
\item{green}{RGB colour code: green}
\item{blue}{RGB colour code: blue}
}
}
\source{
Sharla Gelfand: \url{https://github.com/sharlagelfand/dmc}
}
\usage{
floss
}
\description{
A dataset containing the codes, names and colours of DMC floss colours
}
\keyword{datasets}
