% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_scaled.r
\name{image_to_scaled}
\alias{image_to_scaled}
\title{Scale an image raster array}
\arguments{
\item{image}{A raster array from an image.}

\item{size_unit}{The unit by which the image will be scaled. Can be one of the following options: 'stitches', 'inches', or 'cm'}

\item{img_size}{Size of output image, in the unit determined by size_unit. Use a single value (e.g. \code{48}) for a square image.
Use a vector of two values for a rectangular image \code{c(width, height)}.}

\item{brightness}{A value >1 will increase the brightness of the image while a positive value <1 will decrease the brightness.}

\item{warhol}{Array of values \code{c(1, 2, 3)} associated with R, G, B colour channels. Swap values in array to swap colour channels for a fun visual effect.}
}
\value{
A list with element \code{Img_scaled} containing a data frame of the x- & y-coordinates, R, G, B channels, and hex colour of each stitch (pixel).
}
\description{
Scale the size of an image, in pixel. Process into a data frame.
}
