% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_to_colours.R
\name{scaled_to_colours}
\alias{scaled_to_colours}
\title{Convert image output from image_to_scaled() to cross-stitches}
\arguments{
\item{image_list}{List output from image_to_scaled(). Contains an element  \code{Img_scaled}.}

\item{method}{Default 'cie94'. The method to use for colour comparison. Either 'euclidean', 'cie1976', 'cie94', 'cie2000', or 'cmc'.
See \code{farver::compare_colour}.}

\item{colour_table}{Defaults to \code{dmc_colours}. Data frame of  colours to map onto image. Must contain name and red, green, blue channels.
See attached data  \code{dmc_colours} as examples. Original source for \code{dmc_colours} is: https://github.com/sharlagelfand/dmc}

\item{colour_palette}{Used to supply a character vector of specific DMC colours to use. Can be particularly useful in instances where you already have
many colours available and maintain a list of their DMC codes. Defaults to NULL - in which case all colours are to be used.
Ignored if a \code{colour_table} is supplied.}

\item{trans_bg}{If \code{img} is a png has a transparent background, name of colour to replace the background.}

\item{dithering}{Improves colour of large, photo-realistic mosaics.}

\item{n_colours}{Optional: Integer representing the maximum number of colours to use. This option selects the \code{n_colours} that appear the most, and remaps all other colours to match these \code{n_colours}. Otherwise, defaults to NULL.}
}
\value{
A list with element \code{Img_lego} containing a data frame of the x- & y-coordinates, R, G, B channels, and mapped colour of each brick (pixel).
}
\description{
Match raw colour channel values to a smaller subset of colours.
}
